A Bayesian Heirarchical Random Covariance Model (RCM) for Multi-subject Task-based fMRI Analysis

An efficient heirarchical graphical modeling package for robust Bayesian RCM sampling, diagnostics, inference, and visualization.

Details

The DESCRIPTION file: This package was not yet installed at build time.
Index: This package was not yet installed at build time.
~~ An overview of how to use the package, including the most important ~~ ~~ functions ~~

Author

Quinton Neville [aut, cre]

Maintainer: Q <nevil066@umn.edu>

References

~~ Literature or other references for background information ~~

See also

Examples